feat: refactor layout and middleware for improved internationalization handling
