+++ src/app/middleware.ts
@@ -1,9 +1,22 @@
 import { routing } from '@src/i18n/routing';
 import createMiddleware from 'next-intl/middleware';
+import { NextRequest } from 'next/server';
 
-export default createMiddleware(routing);
+export default async function middleware(request: NextRequest) {
+  //  const defaultLocale = request.headers.get('x-your-custom-locale') || 'en';
+
+  const handleI18nRouting = createMiddleware({
+    locales: routing.locales,
+    defaultLocale: routing.defaultLocale,
+  });
+  const response = handleI18nRouting(request);
+
+  // response.headers.set('x-your-custom-locale', defaultLocale);
+
+  return response;
+}
 
 export const config = {
   // Match only internationalized pathnames
-  matcher: ['/', '/(en-US|fr-CA)/:path*'],
+  matcher: ['/', '/(fr|en)/:path*'],
 };
