+++ tailwind.config.ts
@@ -18,3 +18,5 @@ export default {
     themes: ["light", "dark"],
   },
 } satisfies Config
+
+
